+++ CHANGES.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [upgrade-v14-1]
+- [TASK] Use prefix: q_ as workaround for string values for array keys on selectedAnswers
+
 ## [upgrade-v14-0]
 - [DOCS] Change composer/ext_emconf TYPO3 version to V14
 - [TASK] Migrate Plugin registration/configuration